+++ test/spec/provider/cloud-element-templates/properties/UpdateTemplatePropertiesOrder.spec.js
@@ -79,7 +79,7 @@ describe('provider/cloud-element-templates - UpdateTemplatePropertiesOrder', fun
 
     describe('zeebe:input', function() {
 
-      it('property set', inject(function(elementRegistry, bpmnjs) {
+      it('property set', function() {
 
         // given
         const task = el('TASK');
@@ -96,10 +96,10 @@ describe('provider/cloud-element-templates - UpdateTemplatePropertiesOrder', fun
         ];
 
         expectOrder(ioMapping.inputParameters, expectedInputs);
-      }));
+      });
 
 
-      it('property not set - optional', inject(function(elementRegistry, bpmnjs) {
+      it('property not set - optional', function() {
 
         // given
         const task = el('TASK');
@@ -117,7 +117,7 @@ describe('provider/cloud-element-templates - UpdateTemplatePropertiesOrder', fun
         ];
 
         expectOrder(ioMapping.inputParameters, expectedInputs);
-      }));
+      });
 
 
       it('property not set - conditional', inject(function(modeling) {
@@ -142,12 +142,32 @@ describe('provider/cloud-element-templates - UpdateTemplatePropertiesOrder', fun
         expectOrder(ioMapping.inputParameters, expectedInputs);
       }));
 
+
+      it('properties with matching binding but different conditions', function() {
+
+        // given
+        const task = el('Task2_conditions');
+
+        // when
+        update(task, 'Input 1', 'foo');
+
+        // then
+        const ioMapping = findExtension(task, 'zeebe:IoMapping');
+
+        const expectedInputs = [
+          findInputParameter(ioMapping, binding(task, 'Input 1')),
+          findInputParameter(ioMapping, binding(task, 'Input 2'))
+        ];
+
+        expectOrder(ioMapping.inputParameters, expectedInputs);
+      });
+
     });
 
 
     describe('zeebe:output', function() {
 
-      it('porperty set', inject(function(elementRegistry, bpmnjs) {
+      it('porperty set', function() {
 
         // given
         const task = el('TASK');
@@ -164,10 +184,10 @@ describe('provider/cloud-element-templates - UpdateTemplatePropertiesOrder', fun
         ];
 
         expectOrder(ioMapping.outputParameters, expectedOutputs);
-      }));
+      });
 
 
-      it('porperty not set - optional', inject(function(elementRegistry, bpmnjs) {
+      it('porperty not set - optional', function() {
 
         // given
         const task = el('TASK');
@@ -185,7 +205,7 @@ describe('provider/cloud-element-templates - UpdateTemplatePropertiesOrder', fun
         ];
 
         expectOrder(ioMapping.outputParameters, expectedOutputs);
-      }));
+      });
 
 
       it('porperty not set - conditional', inject(function(modeling) {
@@ -209,12 +229,32 @@ describe('provider/cloud-element-templates - UpdateTemplatePropertiesOrder', fun
         expectOrder(ioMapping.outputParameters, expectedOutputs);
       }));
 
+
+      it('properties with matching binding but different conditions', function() {
+
+        // given
+        const task = el('Task2_conditions');
+
+        // when
+        update(task, 'Output 1', 'foo');
+
+        // then
+        const ioMapping = findExtension(task, 'zeebe:IoMapping');
+
+        const expectedOutputs = [
+          findOutputParameter(ioMapping, binding(task, 'Output 1')),
+          findOutputParameter(ioMapping, binding(task, 'Output 2'))
+        ];
+
+        expectOrder(ioMapping.outputParameters, expectedOutputs);
+      });
+
     });
 
 
     describe('zeebe:property', function() {
 
-      it('property set', inject(function(elementRegistry, bpmnjs) {
+      it('property set', function() {
 
         // given
         const task = el('TASK');
@@ -231,10 +271,10 @@ describe('provider/cloud-element-templates - UpdateTemplatePropertiesOrder', fun
         ];
 
         expectOrder(zeebeProperties.properties, expectedProperties);
-      }));
+      });
 
 
-      it('property not set - optional', inject(function(elementRegistry, bpmnjs) {
+      it('property not set - optional', function() {
 
         // given
         const task = el('TASK');
@@ -252,7 +292,7 @@ describe('provider/cloud-element-templates - UpdateTemplatePropertiesOrder', fun
         ];
 
         expectOrder(zeebeProperties.properties, expectedProperties);
-      }));
+      });
 
 
       it('property not set - conditional', inject(function(modeling) {
@@ -277,12 +317,32 @@ describe('provider/cloud-element-templates - UpdateTemplatePropertiesOrder', fun
         expectOrder(zeebeProperties.properties, expectedProperties);
       }));
 
+
+      it('properties with matching binding but different conditions', function() {
+
+        // given
+        const task = el('Task2_conditions');
+
+        // when
+        update(task, 'Property 1', 'foo');
+
+        // then
+        const zeebeProperties = findExtension(task, 'zeebe:Properties');
+
+        const expectedProperties = [
+          findZeebeProperty(zeebeProperties, binding(task, 'Property 1')),
+          findZeebeProperty(zeebeProperties, binding(task, 'Property 2'))
+        ];
+
+        expectOrder(zeebeProperties.properties, expectedProperties);
+      });
+
     });
 
 
     describe('zeebe:taskHeader', function() {
 
-      it('property set', inject(function(elementRegistry, bpmnjs) {
+      it('property set', function() {
 
         // given
         const task = el('TASK');
@@ -299,7 +359,7 @@ describe('provider/cloud-element-templates - UpdateTemplatePropertiesOrder', fun
         ];
 
         expectOrder(taskHeaders.values, expectedHeaders);
-      }));
+      });
 
 
       it('property not set - conditional', inject(function(modeling) {
@@ -322,6 +382,26 @@ describe('provider/cloud-element-templates - UpdateTemplatePropertiesOrder', fun
 
         expectOrder(taskHeaders.values, expectedHeaders);
       }));
+
+
+      it('properties with matching binding but different conditions', function() {
+
+        // given
+        const task = el('Task2_conditions');
+
+        // when
+        update(task, 'Task Header 1', 'foo');
+
+        // then
+        const taskHeaders = findExtension(task, 'zeebe:TaskHeaders');
+
+        const expectedHeaders = [
+          findTaskHeader(taskHeaders, binding(task, 'Task Header 1')),
+          findTaskHeader(taskHeaders, binding(task, 'Task Header 2'))
+        ];
+
+        expectOrder(taskHeaders.values, expectedHeaders);
+      });
 
 
     });